requirements.txt
numpy==2.2.2
matplotlib==3.10.0
pillow==11.1.0